RM use-file-info-tree.
