+++ src/styles/tiptap.scss
@@ -1,6 +1,5 @@
 .tiptap {
-  // make all text white:
-  color: rgb(244, 244, 244);
+  color: rgb(197, 196, 196);
   height: 100%;
   outline: none;
   padding: 0.5rem;
